update scrape schedule

+++ schedule/scrape_schedule.py
@@ -3,8 +3,9 @@
 ## has been added or removed.
 
 from bs4 import BeautifulSoup
+from django.db.utils import DataError
 import mechanicalsoup
-from datetime import date, datetime
+from datetime import date, datetime, timedelta
 import os, requests, sys
 
 if os.name == 'nt':
@@ -28,6 +29,7 @@ schedule_notes = [] # list that will hold notes if any
 team_events = [] # list that will hold OYHA, OCHL and OWHL teams to merge with oic_schedule[]
 north_locker_rooms = [[1, 3], [2, 4]]  # Locker room numbers in North
 south_locker_rooms = [[6, 9], [5, 8], 7]  # Locker room numbers in South
+
 
 def scrape_oic_schedule(date):
     '''Scrapes Ozaukee Ice Center schedule website the days events.'''
@@ -258,12 +260,13 @@ def add_locker_rooms_to_schedule():
         x += 1
 
 
-def add_schedule_to_model(schedule):
+def add_schedule_to_model(schedule, data_removed):
     '''Adds OIC daily schedule to RinkSchedule model.'''
     model = RinkSchedule
 
-    # First, clear yesterday's schedule from the model
-    RinkSchedule.objects.all().delete()
+    # First, clear the database table once if data_removed = False
+    if not data_removed:
+        RinkSchedule.objects.all().delete()
 
     for item in schedule:
         try:
@@ -286,47 +289,57 @@ if __name__ == "__main__":
     
     the_date = date.today()
     # the_date = "2019-10-26"
-
     scrape_date = date.isoformat(the_date)
-    scrape_oic_schedule(scrape_date)
-    ### UNCOMMENT DURING HOCKEY SEASON ###
-    # Scrape OYHA teams daily
-    try:
-        scrape_oyha_teams(scrape_date)
-    except Exception as e:
-        print(f"{e}, scrape_oyha_teams()")
+    data_removed = False # used to check if the database table has been cleared once
 
-    # If it is Friday, scrape OWHL teams
-    if date.weekday(date.today()) == 4:
-        try:
-            scrape_owhl_teams(scrape_date)
-        except Exception as e:
-            print(f"{e}, scrape_owhl_teams()")
 
-    # If it is Sunday, scrape OCHL teams
-    if date.weekday(date.today()) == 6:
-        try:
-            scrape_ochl_teams()
-        except Exception as e:
-            print(f"{e}, scrape_ochl_teams()")
+    def swap_team_names():
+        ''' Replace schedule event with team names if they match times'''
+        if len(team_events) != 0:
+            for item in team_events:
+                for oic in oic_schedule:
+                    if item[0] == oic[1] and item[3] == oic[3]:
+                        if item[2] == "":
+                            oic[4] = f"{item[1]}"
+                        else:
+                            oic[4] = f"{item[1]} vs {item[2]}"
 
-    # Replace schedule event with team names if they match times
-    if len(team_events) != 0:
-        for item in team_events:
-            for oic in oic_schedule:
-                if item[0] == oic[1] and item[3] == oic[3]:
-                    if item[2] == "":
-                        oic[4] = f"{item[1]}"
-                    else:
-                        oic[4] = f"{item[1]} vs {item[2]}"
+    # If it's not Saturday or Sunday, scrape oic schedule
+    if date.weekday(date.today()) != 5 and date.weekday(date.today()) !=6:
+        scrape_oic_schedule(scrape_date)
 
-    # If anything ends at midnight, change to 11:59 PM
-    # for item in oic_schedule:
-    #     if item[2] == "12:00 AM":
-    #         item[2] = "11:59 PM"
+        # If it is Friday these additional things need to be done
+        if date.weekday(date.today()) == 4:
+            # Scrape OWHL teams
+            try:
+                scrape_owhl_teams(scrape_date)
+            except Exception as e:
+                print(f"{e}, scrape_owhl_teams()")
+        swap_team_names()
+        add_locker_rooms_to_schedule()
+        add_schedule_to_model(oic_schedule, data_removed)
+        data_removed = True
+        oic_schedule.clear()
+        team_events.clear()
 
-    # Insert data into Django model
-    add_locker_rooms_to_schedule()
-    add_schedule_to_model(oic_schedule)
-    # for item in oic_schedule:
-    #     print(item)
+        if date.weekday(date.today()) == 4:
+            saturday = date.isoformat(date.today() + timedelta(days=1))
+            scrape_oic_schedule(saturday)
+            swap_team_names()
+            add_locker_rooms_to_schedule()
+            add_schedule_to_model(oic_schedule, data_removed)
+            oic_schedule.clear()
+            team_events.clear()
+
+            sunday = date.isoformat(date.today() + timedelta(days=2))
+            # oic_schedule.clear()
+            scrape_oic_schedule(sunday)
+            try:
+                scrape_ochl_teams()
+            except Exception as e:
+                print(f"{e}, scrape_ochl_teams()")
+            swap_team_names()
+            add_locker_rooms_to_schedule()
+            add_schedule_to_model(oic_schedule, data_removed)
+            oic_schedule.clear()
+            team_events.clear()
